rename _sync_chat_wrapper sub-module to _chat_wrapper
